Remove current_assignment from jobs
A reassignment will be treated like a new assignment, so there will
never be any returning to old assignments. Therefore, the latest can be
treated as the current one.

A job can go back out for contractors to claim after having a number of
assignments, so its status indicates this (having open)