Multicontractor
This is implemented by allowing multiple jobs to be created with the
same job number. Job number is not the PK, so this is fine. Multiple
jobs with the same job number (which comes from client) essentially
creates a group.

Logic is added to check, on invoicing, whether all the jobs in a job
group are completed. Job numbers only appear once in the invoicing
dialog and all subjobs are updated with the new invoice number.