Allow admin end-to-end job/assignment create
Allow admin to create a job and assignment from start to finish without
contractor intervention.

Admin creates a job, then create an assignment. Adding an actual date at
this point will cause the assignment to be created with "fulfilled"
status already applied (usual validation rules apply about needing a
contractor to be supplied at assignment create)

Usually, when the latest assignment is fulfilled, the parent job goes
into the review state. However, if the assignment was pending (i.e. the
contractor was assigned but didn't acknowledge it) then we
assumes this is an end to end create operation and put the parent job
straight to the completed state, so there's nothing else to do - it can
now be invoiced.